% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{createComputeMode}
\alias{createComputeMode}
\title{createComputeMode}
\usage{
createComputeMode(language = "R", exactBitStream = FALSE,
  extraSort = TRUE, test1 = FALSE, test2 = FALSE, test3 = FALSE)
}
\description{
createComputeMode
}

