% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{fitExample}
\alias{fitExample}
\title{Fit an example DPP model}
\usage{
fitExample(data, n.burn = 10, n.reps = 20, max.row.nbhd.size = 50,
  row.frac.probes = 0.05, col.frac.probes = 0.05, computeMode = "R")
}
\description{
\code{fitExample} fits an example DPP model
}
\examples{
\dontrun{
# Simulate data
simulation <- simulateExample(n = 25, p = 125)

# Fit model
posterior <- fitExample(simulation, n.burn = 10, n.reps = 20)

# Summarize posterior
d_credible.v <- quantile(posterior$d.v, prob=c(.025,.975))
mean.taxicab <- mean(posterior$mean.taxicab.v)
se_mean.taxicab <- sd(posterior$mean.taxicab.v)/sqrt(length(posterior$mean.taxicab.v))
}
}

