% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{npclustData}
\alias{npclustData}
\title{Create a NPClustData object from a dataset}
\usage{
npclustData(raw_data, M = 10, b0 = 2.2, b1 = 20)
}
\arguments{
\item{raw_data}{a \eqn{n * p}{n * p} matrix containing observation values from n subjects on p variables}

\item{M}{The concentration parameter of DP for latent elements}

\item{b0}{The concentration parameter of column-intercept cluster}

\item{b1}{The concentration parameter of PDP for column clustering}
}
\description{
The function \code{npclustData} creates a NPClustData object from the dataset to be analyzed.
}
\details{
The function \code{npclustData} integrates data and prefixed model constants to create a NPClustData object to be passed to \code{npclustFit} for cluster analysis.
}
