% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{npclustFit}
\alias{npclustFit}
\title{Apply Bayesian nonparametric clustering to a dataset}
\usage{
npclustFit(npclustDataObj, n.burn = 10, n.reps = 20,
  max.row.nbhd.size = round(0.1 * 25 * 125^0.5),
  max.col.nbhd.size = round(0.05 * 125), row.frac.probes = 0.05,
  col.frac.probes = 0.1, prob.compute.col.nbhd = 0.2, dahl.flag = FALSE,
  standardize.X = FALSE, flip.sign = FALSE, tBB_flag = FALSE,
  taxicab = FALSE, computeMode = createComputeMode())
}
\arguments{
\item{npclustDataObj}{The NPClustData object containing the data to be analyzed. This can be generated by \code{npclustData} from a real dataset or \code{simulateExample} for a simulated dataset.}

\item{n.burn}{Number of burn-in samples to be generated}

\item{n.reps}{Number of post burn-in samples to be generated}

\item{max.row.nbhd.size}{Maximum neighborhood size for sampling parameters of DP for latent vector elements, should be small compared to \eqn{n*p^d}{n*p^d}}

\item{max.col.nbhd.size}{Maximum neighborhood size for sampling parameters of PDP for variable clustering, should be small compared to \eqn{p}{p}}

\item{row.frac.probes}{The fraction of neighborhood assignment being updated in DP for latent vector elements}

\item{col.frac.probes}{The fraction of neighborhood assignment being updated in PDP for variable clustering}

\item{prob.compute.col.nbhd}{The probability of computing the neighborhood assignment in PDP for variable clustering}

\item{dahl.flag}{Whether or not to compute the least-squares allocation (not included for now)}

\item{standardize.X}{Whether or not to standardize the variables}

\item{flip.sign}{???}

\item{tBB_flag}{Whether or not to compute the \eqn{X^{T} X}{X^T X} for regression}

\item{taxicab}{Whether or not to compute the taxicab distance which measures variable clustering accuracy for simulated datasets}

\item{computeMode}{The computeMode object to be used, generated by \code{createComputeMode}.}
}
\description{
\code{npclustFit} performs Bayesian nonparametric clustering introduced in VariScan model.
}
\examples{
\dontrun{
# Simulate data
simulation <- simulateExample(n = 25, p = 125)

# Fit model
posterior <- npclustFit(simulation, n.burn = 10, n.reps = 20)

# Summarize posterior
d_credible.v <- quantile(posterior$d.v, prob=c(.025,.975))
mean.taxicab <- mean(posterior$mean.taxicab.v)
se_mean.taxicab <- sd(posterior$mean.taxicab.v)/sqrt(length(posterior$mean.taxicab.v))
}

}
