% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{profileExample}
\alias{profileExample}
\title{Profile a simulated example}
\usage{
profileExample(n = 25, p = 250, n.burn = 10, n.reps = 20,
  row.frac.probes = 0.05, col.frac.probes = 0.05,
  computeMode = createComputeMode(), filename = "Rprof.out")
}
\description{
Profile a simulated example
}
