% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{createComputeMode}
\alias{createComputeMode}
\title{createComputeMode}
\usage{
createComputeMode(language = "R", exactBitStream = FALSE,
  extraSort = TRUE, completeTest = FALSE, tolerance = 1e-10,
  specialMode = NULL, test1 = FALSE, test2 = FALSE, test3 = FALSE,
  lowLevelTiming = FALSE)
}
\description{
createComputeMode
}
