% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{npclustFit}
\alias{npclustFit}
\title{Apply Bayesian nonparametric clustering to a dataset}
\usage{
npclustFit(npclustDataObj, n.burn = 10, n.reps = 20,
  max.row.nbhd.size = round(0.1 * 25 * 125^0.5),
  max.col.nbhd.size = round(0.05 * 125), row.frac.probes = 0.05,
  col.frac.probes = 0.1, prob.compute.col.nbhd = 0.2, dahl.flag = FALSE,
  standardize.X = FALSE, flip.sign = FALSE, tBB_flag = FALSE,
  taxicab = FALSE, computeMode = createComputeMode())
}
\arguments{
\item{npclustDataObj}{The NPClustData object containing the data to be analyzed. This can be generated by \code{\link{npclustData}} from a real dataset or \code{\link{simulateExample}} for a simulated dataset.}

\item{n.burn}{Number of burn-in samples to be generated}

\item{n.reps}{Number of post burn-in samples to be generated}

\item{max.row.nbhd.size}{Maximum neighborhood size for sampling parameters of DP for latent vector elements, should be small compared to \eqn{n*p^d}{n*p^d}}

\item{max.col.nbhd.size}{Maximum neighborhood size for sampling parameters of PDP for variable clustering, should be small compared to \eqn{p}{p}}

\item{row.frac.probes}{The fraction of neighborhood assignment being updated in DP for latent vector elements}

\item{col.frac.probes}{The fraction of neighborhood assignment being updated in PDP for variable clustering}

\item{prob.compute.col.nbhd}{The probability of computing the neighborhood assignment in PDP for variable clustering}

\item{dahl.flag}{Whether or not to compute the least-squares allocation (not included for now)}

\item{standardize.X}{Whether or not to standardize the variables}

\item{flip.sign}{A future extension. Currently set to FALSE.}

\item{tBB_flag}{Whether or not to compute the \eqn{X^{T} X}{X^T X} for regression}

\item{taxicab}{Whether or not to compute the taxicab distance which measures variable clustering accuracy for simulated datasets}

\item{computeMode}{The computeMode object to be used, generated by \code{\link{createComputeMode}}.}
}
\value{
An object which is a list of MCMC samples for following:
\item{K.v}{Number of atoms in the PDP base measure from DP}
\item{G.v}{Number of detected clusters among variables}
\item{tau_int.v}{The standard deviation parameter associated with intercept cluster}
\item{tau.v}{The standard deviation parameter for regular variables}
\item{tau_0.v}{The standard deviation parameter for high variation vatiables}
\item{d.v}{The PDP discount parameter}
\item{PDP_log.BF}{The lower bound log-Bayes factor favoring a PDP model instead of DP}
\item{row.flip.v}{The Metropolis-Hastings acceptance rate for sampling DP parameters for latent vector elements}
\item{col_flip.v}{The Metropolis-Hastings acceptance rate for sampling PDP parameters for variable clustering}
\item{col_exit.v}{The proportion for the fast PDP sampling algorithm to propose the exact same allocations as the previous MCMC sample for variables in a nighborhood}
\item{col_new_clust.v}{Proportion of opening a new PDP cluster among all variables}
\item{nbhd_max}{The maximum distance among the neighborhoods in the fast PDP sampling algorithm}
\item{pi.mt}{The pairwise probability matrix of two variables being in the same cluster}
\item{parm}{All parameter values for the last MCMC sample}
\item{init.parm}{All parameter values obtained from initialization}
}
\description{
\code{npclustFit} performs Bayesian nonparametric clustering introduced in VariScan model.
}
\examples{
\dontrun{

## Simulations

# Simulate data
simulation <- simulateExample(n = 25, p = 125)

# Apply the Bayesian nonparametric clustering
posterior <- npclustFit(simulation, n.burn = 10, n.reps = 20, taxicab = T)

# Summarize posterior
d_credible.v <- quantile(posterior$d.v, prob=c(.025,.975)) # The posterior credible interval for PDP discount parameter
mean.taxicab <- mean(posterior$mean.taxicab.v) # The mean taxicab distance
se_mean.taxicab <- sd(posterior$mean.taxicab.v)/sqrt(length(posterior$mean.taxicab.v)) # The standard error of mean taxicab distance

## Real datasets (DLBCL)

# Take a subset of the dataset
raw_data<-as.matrix(dlbcl[sample(1:nrow(dlbcl),size=100),sample(1:ncol(dlbcl),size=500)])

# Convert data to NPClustData object
real_data<-npclustData(raw_data)

# Apply the Bayesian nonparametric clustering
posterior <- npclustFit(real_data, n.burn = 10, n.reps = 20, computeMode = createComputeMode(language = "C"))

# Analyze results
quantile(posterior$d.v, prob=c(.025,.975)) # The posterior credible interval for PDP discount parameter
hist(posterior$G.v) # Histogram of the posterior of number of detected clusters among variables
posterior$parm$clust$c.v # The variable cluster allocations from the last MCMC sample
posterior$parm$clust$C.m.vec # The variable cluster sizes, aligned in order of cluster number, from the last MCMC sample
}

}
\references{
Guha, S., & Baladandayuthapani, V. (2016). A nonparametric bayesian technique for high-dimensional regression. \emph{Electronic Journal of Statistics}, 10(2), 3374-3424. DOI: 10.1214/16-EJS1184
}
