% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{simulateExample}
\alias{simulateExample}
\title{Generates a simulated dataset}
\usage{
simulateExample(n = 25, p = 250, prop.X.miss = 0, tau = 0.5,
  tau_0 = 1.25)
}
\arguments{
\item{n}{Number of subjects}

\item{p}{Number of variables}

\item{prop.X.miss}{Proportion of missing data}

\item{tau}{The variance parameter for regular variables}

\item{tau_0}{The variance paramter for high variation vatiables}
}
\description{
The function \code{simulateExample} generates a NPClustData object from a simulates dataset of desired size
}
